relocate Uri.scheme property to the common module
