+++ module/common/src/iosMain/kotlin/org/cru/godtools/shared/common/model/Uri.ios.kt
@@ -5,6 +5,7 @@ import platform.Foundation.NSURL
 
 @Suppress("CONFLICTING_OVERLOADS")
 actual typealias Uri = NSURL
+actual val Uri.scheme get() = scheme
 
 actual fun String?.toUriOrNull() = this?.let {
     try {
