--- module/parser/src/androidMain/kotlin/org/cru/godtools/shared/tool/parser/util/Uri.android.kt
@@ -1,5 +1,0 @@
-package org.cru.godtools.shared.tool.parser.util
-
-import android.net.Uri
-
-internal actual inline val Uri.scheme get() = scheme
