+++ module/parser/src/commonMain/kotlin/org/cru/godtools/shared/tool/parser/util/Uri.kt
@@ -1,10 +1,6 @@
 package org.cru.godtools.shared.tool.parser.util
 
-import org.cru.godtools.shared.common.model.Uri
 import org.cru.godtools.shared.common.model.toUriOrNull
-
-internal expect val Uri.scheme: String?
-internal val Uri.isHttpUrl: Boolean get() = scheme?.matches(Regex("https?", RegexOption.IGNORE_CASE)) == true
 
 internal fun String?.toAbsoluteUriOrNull(defaultScheme: String = "http") =
     this?.addSchemeIfNecessary(defaultScheme).toUriOrNull()
