--- module/parser/src/iosMain/kotlin/org/cru/godtools/shared/tool/parser/util/Uri.ios.kt
@@ -1,5 +1,0 @@
-package org.cru.godtools.shared.tool.parser.util
-
-import platform.Foundation.NSURL
-
-internal actual inline val NSURL.scheme get() = scheme
